+++ src/components/BurgerMenu/burgerMenu.tsx
@@ -9,7 +9,7 @@ import { TbPoint } from 'react-icons/tb';
 import { Container } from './styles';
 import { useBurgerMenuContext } from '@src/context/burgerMenuContext';
 import Box from '@src/shared/Box/box';
-import Input from '../Navbar/Input/Input';
+import Input from '../Input/Input';
 import Logo from '../Logo/Logo';
 
 export default function BurgerMenu({ isOpen }: { isOpen: boolean }) {
